requirements.txt
Django>=3.2,<3.3.0
djangorestframework>=3.12.4,<3.13.0
pandas>=1.2.4,<1.3.0
scikit-learn>=0.24.2,<0.25.0